+++ src/Sutil/Program.fs
@@ -1,13 +1,9 @@
 namespace Sutil
 
 open Core
-open CoreElements
 open Browser.Types
-open Browser.Dom
 
-/// <summary>
-/// Main entry points for a Sutil program
-/// </summary>
+/// <exclude/>
 module Program =
     open System
 
@@ -17,31 +13,39 @@ module Program =
 
     // -----------------------------------------------------------------------------------------------------------------------------
 
-    [<Obsolete("Use Program.mount( host, app )")>]
+    [<Obsolete("Use Program.mount")>]
     let mountElementOnDocumentElement (host : HTMLElement) (app : SutilElement)  =
         app |> _mount (MountOp.AppendTo,ElementRef.Element host)
 
-    [<Obsolete("Use Program.mount( doc, id, app )")>]
+    [<Obsolete("Use Program.mount")>]
     let mountElementOnDocument (doc : Document) id (app : SutilElement)  =
         let host = doc.querySelector($"#{id}") :?> HTMLElement
         //mountElementOnDocumentElement host app
         app |> _mount (MountOp.AppendTo,ElementRef.Element host)
 
-    [<Obsolete("Use Program.mount( host, app )")>]
+    [<Obsolete("Use Program.mount")>]
     let mountDomElement (host : HTMLElement) (app : SutilElement)  =
         //mountElementOnDocumentElement host app
         app |> _mount (MountOp.AppendTo,ElementRef.Element host)
 
-    [<Obsolete("Use Program.mount( id, app )")>]
+    [<Obsolete("Use Program.mount")>]
     let mountElement id (app : SutilElement)  =
         //mountElementOnDocument document id (exclusive app)
         app |> _mount (MountOp.AppendTo,ElementRef.Id id)
 
-    [<Obsolete("Use Program.mountAfter( host, app )")>]
+    [<Obsolete("Use Program.mountAfter")>]
     let mountElementAfter (prev : HTMLElement) (app : SutilElement) =
         app |> _mount (MountOp.InsertAfter,ElementRef.Element prev)
 
-
+/// <summary>
+/// Main entry points for a Sutil program
+/// </summary>
+/// <example>
+/// For example, this will mount the <c>SutilElement</c> returned from <c>view()</c> onto the <c>div</c> with id "sutil-app"
+/// <code>
+/// view () |> Program.mount
+/// </code>
+/// </example>
 type Program() =
     ///<summary>
     /// Mount application on element with given id.